Update tests due to changes
