+++ test/base/CWToken.test.ts
@@ -3,7 +3,6 @@ import { expect } from "chai";
 import { Contract, ContractFactory } from "ethers";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
 import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
-import { proveTx } from "../test-utils/eth";
 
 async function setUpFixture<T>(func: () => Promise<T>): Promise<T> {
   if (network.name === "hardhat") {
@@ -13,21 +12,19 @@ async function setUpFixture<T>(func: () => Promise<T>): Promise<T> {
   }
 }
 
-describe("Contract 'BRLCToken'", async () => {
-  const TOKEN_NAME = "BRL Coin";
-  const TOKEN_SYMBOL = "BRLC";
+describe("Contract 'CWToken'", async () => {
+  const TOKEN_NAME = "CW Token";
+  const TOKEN_SYMBOL = "CWT";
   const TOKEN_DECIMALS = 6;
 
   const REVERT_MESSAGE_INITIALIZABLE_CONTRACT_IS_ALREADY_INITIALIZED = "Initializable: contract is already initialized";
-  const REVERT_MESSAGE_INSUFFICIENT_ALLOWANCE = "ERC20: insufficient allowance";
 
   let tokenFactory: ContractFactory;
   let deployer: SignerWithAddress;
-  let user: SignerWithAddress;
 
   before(async () => {
-    [deployer, user] = await ethers.getSigners();
-    tokenFactory = await ethers.getContractFactory("BRLCToken");
+    [deployer] = await ethers.getSigners();
+    tokenFactory = await ethers.getContractFactory("CWToken");
   });
 
   async function deployToken(): Promise<{ token: Contract }> {
@@ -46,7 +43,6 @@ describe("Contract 'BRLCToken'", async () => {
       expect(await token.pauser()).to.equal(ethers.constants.AddressZero);
       expect(await token.rescuer()).to.equal(ethers.constants.AddressZero);
       expect(await token.mainBlocklister()).to.equal(ethers.constants.AddressZero);
-      expect(await token.mainMinter()).to.equal(ethers.constants.AddressZero);
     });
 
     it("Is reverted if called for the second time", async () => {
@@ -62,35 +58,6 @@ describe("Contract 'BRLCToken'", async () => {
       await expect(
         tokenImplementation.initialize(TOKEN_NAME, TOKEN_SYMBOL)
       ).to.be.revertedWith(REVERT_MESSAGE_INITIALIZABLE_CONTRACT_IS_ALREADY_INITIALIZED);
-    });
-  });
-
-  describe("Function 'isBRLCoin'", async () => {
-    it("Returns true", async () => {
-      const { token } = await setUpFixture(deployToken);
-      expect(await token.isBRLCoin()).to.eq(true);
-    });
-  });
-
-  describe("Function 'transferFrom()'", async () => {
-    it("Executes as expected for non-trusted and trusted accounts", async () => {
-      const maxAmount = ethers.constants.MaxUint256;
-      const userBalance = 123;
-
-      const { token } = await setUpFixture(deployToken);
-      await proveTx(token.updateMainMinter(deployer.address));
-      await proveTx(token.configureMinter(deployer.address, maxAmount));
-      await proveTx(token.mint(user.address, userBalance));
-
-      await expect(
-        token.connect(deployer).transferFrom(user.address, deployer.address, userBalance)
-      ).to.be.revertedWith(REVERT_MESSAGE_INSUFFICIENT_ALLOWANCE);
-
-      await proveTx(token.configureTrustedAccount(deployer.address, true));
-
-      await expect(
-        token.connect(deployer).transferFrom(user.address, deployer.address, userBalance)
-      ).to.be.changeTokenBalances(token, [user, deployer], [-userBalance, +userBalance]);
     });
   });
 });
